Update user permissions and CloudUser serialization
Adjusted user permissions to prevent users from deleting their own accounts. Introduced CloudUser serialization and adjusted the registration process to create an associated CloudUser instance for storing additional data. These changes improve both security and functionality in user management and authentication.

+++ authentication/views.py
@@ -51,7 +51,11 @@ class UsersViewSet(
     ordering = ["username"]
     permission_classes = [IsAuthenticated, IsAdminUser]
 
+    # Мне нужно запретить пользователю удалять самого себя и выдавать ошибку при попытке удалить другого пользователя
     def perform_destroy(self, instance):
+        if self.request.user is instance:
+            return
+
         for file in File.objects.filter(user_id=instance.id):
             delete_file(file.file_path)
